requirements.txt
python 3.9.13
click-8.1.3
colorama-0.4.5
joblib-1.1.0
nltk-3.7
regex-2022.8.17
tqdm-4.64.1
charset-normalizer-2.1.1
idna-3.3
requests-2.28.1
urllib3-1.26.12
filelock-3.8.0
huggingface-hub-0.9.1
pyyaml-6.0
tokenizers-0.12.1
transformers-4.22.1
torch-1.12.1
MarkupSafe-2.1.1
blis-0.9.1
catalogue-2.0.8
confection-0.0.1
cymem-2.0.6
jinja2-3.1.2
langcodes-3.3.0
murmurhash-1.0.8
numpy-1.23.3
packaging-21.3
pathy-0.6.2
preshed-3.0.7
pydantic-1.9.2
pyparsing-3.0.9
smart-op
en-5.2.1
spacy-3.4.1
spacy-legacy-3.0.10
spacy-loggers-1.0.3
srsly-2.4.4
thinc-8.1.1
typer-0.4.2
typing-extensions-4.3.0
wasabi-0.10.1
